Make sure that updates are recorded when calculate is submitted
WPeC's totally awesome checkout page shipping calculator has a submit button that will send some of the shipping data to us in an AJAX request.  The format of the data as of version 3.8.14.1 uses the 'collected_data' array format just like in checkout. We should process this array in case it has some updates to the user meta (checkout information) that haven't been recorded at the time the calculate button was clicked.