Utility functions to help with consistent branding of WP eCommerce
Helper function returning the properly formatted name of WP eCommerce as text

Helper function returning the properly formatted name of WP eCommerce as html for display on in the product, plugin page or elsewhere.

Helper function echoing the properly formatted name of WP eCommerce as html for display on a plugin page or elsewhere.