Running a synchronous http transaction in the main javascript thread is now deprecated in Chrome IE and based on somewhat reliable internet conversations Firefox.
Even though we run our very light weight transaction off of a timeout trigger, it is being detected as part of the main thread.

So we need to make an adjustment to stop things like add to cart, update shipping and checkout from breaking.

Change the checkVisitorId() function to use an asynchronous ajax transaction rather than a synchronous transaction, that's easy. But that still leaves the fringe possibility where if a user makes a first visit to a WP-eCommerce site, the site is using some type of cache ( that means server cache like WP-SuperCache or W3-Total-Cache or a CDN like Akamai or Cloudflare ), and the first page is a page that uses ajax the customer cookie may not be set properly.

Conflicts:
	wpsc-core/js/wp-e-commerce.js